requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.10.0
python-dotenv==1.0.0
ortools>=9.12.0
PyJWT[crypto]==2.10.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
requests==2.31.0
httpx==0.27.0
websockets>=13.0
pytest==7.4.3
gunicorn==21.2.0